Use -Wextra when building C files

+++ src/test/protect_rr_fds.c
@@ -20,7 +20,7 @@ int main(int argc, char* argv[]) {
      those don't interfere with rr by closing RR_RESERVED_ROOT_DIR_FD
      or some other essential file descriptor. */
   test_assert(0 == getrlimit(RLIMIT_NOFILE, &nofile));
-  for (fd = STDERR_FILENO + 1; fd < nofile.rlim_cur; ++fd) {
+  for (fd = STDERR_FILENO + 1; fd < (int)nofile.rlim_cur; ++fd) {
     ret = fcntl(fd, F_SETFD, FD_CLOEXEC);
     test_assert(ret == 0 || (ret == -1 && errno == EBADF));
     ret = dup2(STDERR_FILENO, fd);
